# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EcoHome Store API",
    "version": "1.0.0-beta",
    "description": "API REST para EcoHome Store con JWT, refresh tokens, RBAC y productos con paginacion opcional."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "tags": [
    { "name": "Health" },
    { "name": "Auth" },
    { "name": "Products" },
    { "name": "Users" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "required": ["status", "message"],
        "properties": {
          "status": { "type": "string", "example": "ok" },
          "message": { "type": "string", "example": "EcoHome Store API is running" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string", "example": "Invalid credentials" }
        }
      },
      "ValidationErrorItem": {
        "type": "object",
        "required": ["field", "message"],
        "properties": {
          "field": { "type": "string", "example": "email" },
          "message": { "type": "string", "example": "A valid email is required" }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "required": ["message", "errors"],
        "properties": {
          "message": { "type": "string", "example": "Validation failed" },
          "errors": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ValidationErrorItem" }
          }
        }
      },
      "User": {
        "type": "object",
        "required": ["id", "name", "email", "role", "created_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid", "example": "5c2da17d-3d6e-4c15-aa0a-8f4d4fe9ec59" },
          "name": { "type": "string", "example": "Ana Perez" },
          "email": { "type": "string", "format": "email", "example": "ana@example.com" },
          "role": { "type": "string", "enum": ["admin", "staff", "client"], "example": "client" },
          "created_at": { "type": "string", "format": "date-time", "example": "2026-04-07T18:30:00.000Z" }
        }
      },
      "SignupResponse": {
        "type": "object",
        "required": ["message", "user", "token"],
        "properties": {
          "message": { "type": "string", "example": "User created successfully" },
          "user": { "$ref": "#/components/schemas/User" },
          "token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": ["message", "user", "token", "accessToken", "refreshToken"],
        "properties": {
          "message": { "type": "string", "example": "Login successful" },
          "user": { "$ref": "#/components/schemas/User" },
          "token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.access" },
          "accessToken": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.access" },
          "refreshToken": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh" }
        }
      },
      "SignupRequest": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": { "type": "string", "example": " Rafael Ricardo Mantilla" },
          "email": { "type": "string", "format": "email", "example": "ricardo.mantilla@uniasturias.com" },
          "password": { "type": "string", "format": "password", "minLength": 8, "example": "SecurePass123" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "ricardo.mantilla@uniasturias.com" },
          "password": { "type": "string", "format": "password", "example": "SecurePass123" }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh" }
        }
      },
      "RefreshResponse": {
        "type": "object",
        "required": ["accessToken"],
        "properties": {
          "accessToken": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.access" }
        }
      },
      "Product": {
        "type": "object",
        "required": ["id", "name", "price", "created_at", "updated_at", "created_by"],
        "properties": {
          "id": { "type": "string", "format": "uuid", "example": "a0d13a1a-3f20-4d92-9b93-7aa94a8fcd22" },
          "name": { "type": "string", "example": "Lampara de bambu" },
          "price": { "type": "number", "format": "float", "example": 89.9 },
          "created_by": { "type": "string", "format": "uuid", "nullable": true, "example": "5c2da17d-3d6e-4c15-aa0a-8f4d4fe9ec59" },
          "creator": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": { "type": "string", "format": "uuid", "example": "5c2da17d-3d6e-4c15-aa0a-8f4d4fe9ec59" },
              "name": { "type": "string", "example": "Ana Perez" }
            }
          },
          "created_at": { "type": "string", "format": "date-time", "example": "2026-04-07T18:30:00.000Z" },
          "updated_at": { "type": "string", "format": "date-time", "example": "2026-04-07T18:30:00.000Z" }
        }
      },
      "UserStats": {
        "type": "object",
        "required": ["userId", "name", "productCount"],
        "properties": {
          "userId": { "type": "string", "format": "uuid", "example": "5c2da17d-3d6e-4c15-aa0a-8f4d4fe9ec59" },
          "name": { "type": "string", "example": "Ana Perez" },
          "productCount": { "type": "integer", "example": 4 }
        }
      },
      "CreateProductRequest": {
        "type": "object",
        "required": ["name", "price"],
        "properties": {
          "name": { "type": "string", "example": "Lampara de bambu" },
          "price": { "type": "number", "format": "float", "example": 89.9 }
        }
      },
      "UpdateProductRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "Lampara de bambu premium" },
          "price": { "type": "number", "format": "float", "example": 99.9 }
        }
      },
      "DeleteProductResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string", "example": "Product deleted successfully" }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "required": ["page", "limit", "total"],
        "properties": {
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 10 },
          "total": { "type": "integer", "example": 45 }
        }
      },
      "PaginatedProductsResponse": {
        "type": "object",
        "required": ["data", "pagination"],
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Product" }
          },
          "pagination": { "$ref": "#/components/schemas/PaginationMeta" }
        }
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "tags": ["Health"],
        "operationId": "getHealth",
        "summary": "Check API health",
        "responses": {
          "200": {
            "description": "API available",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HealthResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "operationId": "signup",
        "summary": "Register a client user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SignupRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SignupResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "operationId": "login",
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "operationId": "refreshAccessToken",
        "summary": "Issue a new access token using a refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access token issued",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RefreshResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": ["Products"],
        "operationId": "getProducts",
        "summary": "List products with optional pagination and filtering",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1 },
            "description": "Page number. If omitted together with other query params, the endpoint returns the legacy array response."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "maximum": 100 },
            "description": "Page size."
          },
          {
            "name": "created_by",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "uuid" },
            "description": "Filter products by creator user id."
          }
        ],
        "responses": {
          "200": {
            "description": "Product list. Returns either the legacy array format or a paginated object depending on query params.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Product" }
                    },
                    { "$ref": "#/components/schemas/PaginatedProductsResponse" }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "operationId": "createProduct",
        "summary": "Create product (admin or staff)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProductRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Product" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": ["Products"],
        "operationId": "getProductById",
        "summary": "Get product by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Product" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Products"],
        "operationId": "updateProduct",
        "summary": "Partially update product (admin or staff)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProductRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Product" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Products"],
        "operationId": "replaceProduct",
        "summary": "Update product using PUT (admin or staff)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProductRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Product" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Products"],
        "operationId": "deleteProduct",
        "summary": "Delete product (admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteProductResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/users/me/stats": {
      "get": {
        "tags": ["Users"],
        "operationId": "getMyStats",
        "summary": "Get current user product metrics (admin or staff)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User stats",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserStats" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  }
}